Delete user makes sure that User ID exists

+++ restdb.go
@@ -52,6 +52,12 @@ func DeleteUser(ID int) bool {
 	}
 	defer db.Close()
 
+	t := FindUserID(ID)
+	if t.ID = 0 {
+		log.Println("User does not exist.")
+		return false
+	}
+
 	stmt, err := db.Prepare("DELETE FROM users WHERE ID = $1")
 	if err != nil {
 		log.Println("DeleteUser:", err)
